+++ src/data.c
@@ -29,7 +29,7 @@ void skip_record(FILE *file) {
 // Reading record into memory
 Record *mem_record(FILE *file) {
     unsigned char *str1 = NULL, c = fgetc(file);
-    int size = 0;
+    size_t size = 0;
     Record *result = init_record();
     while (c < 160) {
         unsigned char *tmp = str1;
@@ -61,9 +61,9 @@ Record *mem_record(FILE *file) {
 }
 
 // Counting records in one file
-int count_records(FILE *file) {
+size_t count_records(FILE *file) {
     fseek(file, 0, SEEK_SET);
-    int result = 0;
+    size_t result = 0;
     unsigned char c = fgetc(file);
     c = fgetc(file);
     if (file) {  // ! In case if this file will not be opened
@@ -77,9 +77,9 @@ int count_records(FILE *file) {
 }
 
 // Counting all the records in all .txt files
-int count_all_records(char **file_names, const int files_count) {
-    int result = 0;
-    for (int i = 0; i < files_count; i++) {
+size_t count_all_records(char **file_names, const size_t files_count) {
+    size_t result = 0;
+    for (size_t i = 0; i < files_count; i++) {
         FILE *file = fopen(file_names[i], "r");
         if (file) {
             result += count_records(file);
@@ -90,10 +90,10 @@ int count_all_records(char **file_names, const int files_count) {
 }
 
 // Input one record from 
-Record *read_record(FILE *file, int pos) {
+Record *read_record(FILE *file, size_t pos) {
     fseek(file, 0, SEEK_SET);
     Record *record = NULL;
-    for (int i = 1; i < pos; i++)
+    for (size_t i = 1; i < pos; i++)
         skip_record(file);
     record = mem_record(file);
     return record;
